final commit

+++ src/Pages/MyToys/MyToys.jsx
@@ -2,43 +2,77 @@ import useTitle from '../../Routes/useTitle';
 import { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../../Providers/AuthProviders';
 import { Link } from 'react-router-dom'
+import Swal from 'sweetalert2';
 
 const MyToys = () => {
     useTitle('My-Toys');
-
+    const [isSort, setSort] = useState(true)
     const { user } = useContext(AuthContext);
-
     const [myData, setMyData] = useState([]);
 
     useEffect(() => {
-        fetch(`https://kidolls-server.vercel.app/myToys?email=${user?.email}`)
+            fetch(`https://kidolls-server.vercel.app/myToys?email=${user?.email}&isSort=${isSort}`)
             .then(res => res.json())
             .then(data => { setMyData(data) })
 
-    }, [user.email])
+    }, [user.email, isSort])
 
     console.log(user.email);
     console.log(myData);
 
     const handleDelete = (id) => {
         console.log(id)
-        fetch(`https://kidolls-server.vercel.app/myToys/${id}`, {
-            method: "DELETE",
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://kidolls-server.vercel.app/myToys/${id}`, {
+                    method: "DELETE",
+                })
+                    .then(res => res.json())
+                    .then(result => {
+                        if (result.deletedCount > 0) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your file has been deleted.',
+                                'success'
+                            )
+
+                            fetch(`https://kidolls-server.vercel.app/myToys?email=${user?.email}`)
+                                .then(res => res.json())
+                                .then(data => { setMyData(data) })
+                        }
+                        console.log(result)
+                    })
+
+            }
         })
-            .then(res => res.json())
-            .then(result => {
-                if (result.deletedCount > 0) {
-                    fetch(`https://kidolls-server.vercel.app/myToys?email=${user?.email}`)
-                        .then(res => res.json())
-                        .then(data => { setMyData(data) })
-                }
-                console.log(result)
-            })
+
+
+
+
+
+
     }
+
+
 
     return (
         <div>
             <div className="overflow-x-auto">
+                {
+
+                    isSort ? <button className="p-2 text-green-600 text-lg mb-2 bg-red-100 rounded-lg " onMouseUp={() => setSort(!isSort)}> Low To High </button> :
+                        <button className="p-2 text-green-600 text-lg mb-2 bg-red-100 rounded-lg " onMouseUp={() => setSort(!isSort)} > High To Low </button>
+
+                }
                 <table className="table table-compact w-full">
                     <thead>
                         <tr>
